Pass `request_encoding_type` into `DTOInterface.on_registration()`.
The prior behavior of accessing `route_handler.parsed_fn_signature.parameters["data"]`
from within `on_registration()` fails for the case where the handler is a websocket
handler - as the `"data"` param is in the `websocket_listener` handler signature, not
the `WebsocketHandler` signature.
