+++ litestar/dto/factory/utils.py
@@ -7,19 +7,15 @@ from typing import TYPE_CHECKING, TypeVar
 from msgspec import Struct
 from typing_extensions import get_type_hints
 
-from litestar.params import BodyKwarg
-from litestar.utils.signature import ParsedParameter, ParsedType
+from litestar.utils.signature import ParsedType
 
 from .config import DTOConfig
 
 if TYPE_CHECKING:
     from typing import Any
 
-    from litestar.enums import RequestEncodingType
-
 __all__ = (
     "get_model_type_hints",
-    "infer_request_encoding_from_parameter",
     "parse_configs_from_annotation",
 )
 
@@ -55,21 +51,3 @@ def parse_configs_from_annotation(parsed_type: ParsedType) -> tuple[DTOConfig, .
         The type and config object extracted from the annotation.
     """
     return tuple(item for item in parsed_type.metadata if isinstance(item, DTOConfig))
-
-
-def infer_request_encoding_from_parameter(param: ParsedParameter) -> RequestEncodingType | str | None:
-    """Infer the request encoding type from a parsed type.
-
-    Args:
-        param: The parsed parameter to infer the request encoding type from.
-
-    Returns:
-        The inferred request encoding type.
-    """
-    if param.has_default and isinstance(param.default, BodyKwarg):
-        return param.default.media_type
-    if param.parsed_type.metadata:
-        for item in param.parsed_type.metadata:
-            if isinstance(item, BodyKwarg):
-                return item.media_type
-    return None
